benchmark/diffusion/text-to-image/requirements.txt
torch==2.2.1
diffusers==0.26.3
datasets
transformers
pillow
pynvml
zeus-ml
